+++ src/video/SpriteChecker.hh
@@ -5,8 +5,9 @@
 #include "VDPVRAM.hh"
 #include "VRAMObserver.hh"
 #include "DisplayMode.hh"
-#include "serialize_meta.hh"
+#include "narrow.hh"
 #include "ranges.hh"
+#include "serialize_meta.hh"
 #include "unreachable.hh"
 #include <array>
 #include <cstdint>
@@ -102,8 +103,8 @@ public:
 		// The following is only required when switching from sprite
 		// mode0 to some other mode (in other case it has no effect).
 		// Because in mode 0, currentLine is not updated.
-		currentLine = frameStartTime.getTicksTill_fast(time)
-		                 / VDP::TICKS_PER_LINE;
+		currentLine = narrow<int>(frameStartTime.getTicksTill_fast(time)
+		                 / VDP::TICKS_PER_LINE);
 		// Every line in mode0 has 0 sprites, but none of the lines
 		// are ever requested by the renderer, except for the last
 		// line, because sprites are checked one line before they
@@ -174,8 +175,8 @@ public:
 		// the display line. In reality, sprite checking is probably done
 		// during most of the line. Run tests on real MSX to make a more
 		// accurate model of sprite checking.
-		int limit = frameStartTime.getTicksTill_fast(time)
-		               / VDP::TICKS_PER_LINE;
+		int limit = narrow<int>(frameStartTime.getTicksTill_fast(time)
+		               / VDP::TICKS_PER_LINE);
 		if (currentLine < limit) {
 			// Call the right update method for the current display mode.
 			(this->*updateSpritesMethod)(limit);
